experiments/color_generation/requirements.txt
# Core ML dependencies
torch>=1.9.0
transformers>=4.21.0
numpy>=1.21.0
accelerate>=0.20.0

# Analysis and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
click>=8.0.0

# Development and testing
pytest>=7.0.0